feat: separate core logic from cli
