+++ Cargo.toml
@@ -1,23 +1,3 @@
-[package]
-name = "acevo_content_editor"
-version = "0.1.6"
-edition = "2021"
-description = "A utility tool to quickly unpack Assetto Corsa Evo packed content packages (*.kspkg)."
-
-[dependencies]
-bytemuck = { version = "1.21.0", features = [
-    "derive",
-    "min_const_generics",
-    "must_cast",
-] }
-clap = { version = "4.5.27", features = ["derive", "env"] }
-colored = "3.0.0"
-glob = "0.3.2"
-humansize = "2.1.3"
-humantime = "2.1.0"
-spdlog-rs = { version = "0.4.0", features = ["multi-thread"] }
-thiserror = "2.0.11"
-tokio = { version = "1.43.0", features = ["full"] }
-
-[build-dependencies]
-bindgen = "0.71.1"
+[workspace]
+resolver = "3"
+members = ["cli", "core"]
