+++ cli/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "acevo_content_editor_cli"
+version = "0.1.0"
+edition = "2021"
+description = "A CLI utility tool to quickly unpack Assetto Corsa Evo packed content packages (*.kspkg)."
+
+[dependencies]
+acevo_content_editor = { path = "../core", features = ["async"] }
+clap = { version = "4.5.27", features = ["derive", "env"] }
+glob = "0.3.2"
+spdlog-rs = { version = "0.4.0", features = ["multi-thread"] }
+tokio = { version = "1.43.0", features = ["full"] }
+colored = "3.0.0"
+humansize = "2.1.3"
+humantime = "2.1.0"
