+++ cli/src/commands/list/mod.rs
