+++ cli/src/functions/logging.rs
