+++ cli/src/functions/utility.rs
