+++ core/src/lib.rs
@@ -4,7 +4,5 @@
 
 include!("bindings/bindings.rs");
 
-pub mod args;
-pub mod functions;
 pub mod models;
 pub mod traits;
