+++ core/src/models/file_entry.rs
