+++ core/src/models/mod.rs
@@ -1,11 +1,9 @@
-mod commands;
 mod file_entry;
 mod package_file_table;
 mod packed_package_buffer;
 mod path_tree;
 mod benchmarked;
 
-pub use commands::*;
 pub use file_entry::*;
 pub use package_file_table::*;
 pub use packed_package_buffer::*;
